[Schedule.API] Correcion de errores, cleanup #35

+++ Schedule.API/Models/Repositories/ISeccionesRepository.cs
@@ -7,6 +7,6 @@ namespace Schedule.API.Models.Repositories
     {
         IEnumerable<SeccionesDetailsDTO> GetAllCurrent();
 
-        SeccionesDetailsDTO GetCurrent(int codigo);
+        SeccionesDetailsDTO GetCurrent(ushort codigo);
     }
 }
